requirements.txt
opencv-contrib-python
opencv-python
roboticstoolbox-python==1.0.3
coppeliasim-zmqremoteapi-client
commentjson
pandas